Add a Meson build system
This allows bwrap to be built as a subproject in larger Meson projects.
When built as a subproject, we install into the --libexecdir and require
a program prefix to be specified: for example, Flatpak would use
program_prefix=flatpak- to get /usr/libexec/flatpak-bwrap.

One feature that is deliberately missing is an equivalent of
--with-priv-mode=setuid. On distributions like Debian <= 10 that still
require a setuid bwrap, the sysadmin or distribution packaging will need
to set the correct permissions on the bwrap executable; Debian already
did this in the packaging rather than the upstream build system.

Loosely based on previous work by Jussi Pakkanen (see #133).

Signed-off-by: Simon McVittie <smcv@collabora.com>